Add new example for chapter 5
A new directory /examples is added for example codes. .clang-format is
copied from sysprog21/lkmpg project.

This commit provides a simplified implementation of thread pool. After
initializing the thread pool with thread count, jobs can be added. The
job queue is a SPMC ring buffer. To keep the implementation minimal,
the producer is not protecting, resulting the thread pool can not run
automatically when jobs are added. Or the worker may try to get the job
before it is fully enqueued.

Padding is added in thread_pool_t to avoid false sharing. The number
"40" is the sum of size of struct members, including alignment, before
the first padding. There should be a better way to determine the value
since structure packing is implementation defined.

The test in main function results nondeterminate order of jobs echoing
its id. A mechanism to wait all jobs to complete should be added later
instead of using sleep. `thread_pool_destroyed()` is not functional yet.