Add use case of atomic exhange, TAS, fetch AND
Atomic flag is used for checking if the given thread pool has been
initialized. The flag is initialized when thread pool struct declared and
reseted to false when thread pool destroyed.

Atomic exchange obtains previous state when destroying thread pool and gives
warnning message if the state were running.

Atomic fetch and AND with zero demonstrate a way to set state to idle.