Revise section 5 with new example
Clarification on read-modify-write is first added at the beginning of
the section. To make the discussion based on atomic load/store, more
information is supplemented at the end of section 2.

Example code is included by using minted package. Each subsection is
revised according to the atomic library usage in the example.

At the end of sectoin 5, a new subsection "further improvements" is
added to discuss topics on leveraging other memory order, false sharing
and safe memory reclamation. First two topics are forward referenced to
according chapters. The last one is not covered in this book so it has no
reference.