Add race conditoin example
A diff file is added to patch original example to the one that can cause
races. Substituting `threads.h` with `pthread.h` is also included in
diff because sanitizer hasn't support C11 thread yet. How sanitizer
works and how to use it are added as well. Then explainations on
warning messages from Tsan are followed.

The part mentioning safe memory reclaimation is moved to this subsection
because warning messages from Tsan mentioned it.

A missing reference to spinlock (originally as one of the rmw examples)
in section 9 was added back as a new code block.