Refine the statement on atomicity
The original statement is only true on the successful operation. To 
other failed operations, it is the successful one that finished before 
failed ones.

The new statement does not totally cover the charateristic of atomic 
operations though. It is the generated cmpxchg or LL/SC loop that make 
the operation keeps trying and eventually finish. But considering that 
the purpose of this paragraph is to plot a big picture of order and 
atomicity, more details on atomic operations should be covered in 
section "Atomic as building blocks". More refences to compiler and CPU 
menufacturer documents should be taken in consideration then.