Update descriptions referencing to section 5
The intro of section 10 originally references back to spinlock in 
section 5. It is now replace by using the new example.

Section 10.2 onriginally  references back to UI thread in section 5. It 
is now removed as new example presents down below to explain relaxed 
memory model. New example is used here as well is because it is exactly 
what original example was talking about.