Remove the thread santizer usage and diff file
Introducing thread sanitizer here may be an unexpected pop up for the
readers that are new to concurrency. Here we focus on rmw atomic
operation instead, thus related content and diff file are removed.
The proper place for this topic could be a dedicated section for
"testing, debugging and verifing concurrent programs"

This aligns the decision sticking to C11 thread as well.


Co-authored-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>