Add atomic instruction and simplify rmw_example.c
Add the description of atomic instruction to let readers know there is a
difference between using fetch and..., which is only a programming tool,
and its actual execution as an atomic operation that depends on the
compiler.

Simplify the rmw_example code to provide more flexible examples.

 - Initially, all worker threads will be initialized. The main thread
will ask all workers to start running. If there is no job or the job is
completed, the worker will become idle. Next, the main thread will
continue to add more jobs and ask the worker to start running again.
Meanwhile, the main thread will also wait for the results of the work.

 - Use the struct `tpool_future` to record all the information required
for the job.

Co-authored-by: Chih-Wei Chien <idoleat@taiker.tw>